+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Edge
 
+Minor Changes:
+
+ - `Viking.Model#set` now changes the prototype of the model to type if type was passed
+ - `Viking.Model#set` initializes new `hasMany` associations when set is called with new type
+ 
 Bugfixes:
 
   - `Router#start` now takes `options` and passes them to `Backbone.history.start`